Make into a package
Allows use of entry points to define our executable.
Don't keep re-instancing cursor and connection objects, re-use the ones
instanced in the object constructor.  Set the initial version of the
package to 0.0.3, assume that 0.0.1 was what was present when the
project was left off in spring of 2014.  v0.0.2 is when we got it
working again on hub zero.